backend/requirements.txt
fastapi==0.115.12
openai==1.77.0
PyMuPDF==1.26.0
python-dotenv==1.0.1
python-multipart==0.0.9
uvicorn[standard]==0.34.2
